wire up basic social context link store
