--- src/languages/junto-public-social-context/dna.js
@@ -1,4 +1,0 @@
-import dna from './social-context.dna.gz'
-
-export const DNA = Buffer.from(dna, 'base64')
-export const DNA_NICK = "socialcontext"
