+++ src/languages/social-context/tsconfig.json
